machine-learning-client/requirements.txt
pymongo
openai-whisper
flask
python-dotenv